Refactor ToolChains into separate files. (#16091)
Format the code and factor together some common functionality at the same time.
