+++ include/swift/Driver/ToolChain.h
@@ -15,7 +15,9 @@
 
 #include "swift/Basic/LLVM.h"
 #include "swift/Driver/Action.h"
+#include "swift/Driver/Job.h"
 #include "swift/Frontend/FileTypes.h"
+#include "swift/Option/Options.h"
 #include "llvm/ADT/Triple.h"
 #include "llvm/Option/Option.h"
 
@@ -23,11 +25,11 @@
 
 namespace swift {
 namespace driver {
-  class CommandOutput;
-  class Compilation;
-  class Driver;
-  class Job;
-  class OutputInfo;
+class CommandOutput;
+class Compilation;
+class Driver;
+class Job;
+class OutputInfo;
 
 /// A ToolChain is responsible for turning abstract Actions into concrete,
 /// runnable Jobs.
@@ -47,7 +49,7 @@ protected:
   ToolChain(const Driver &D, const llvm::Triple &T) : D(D), Triple(T) {}
 
   /// A special name used to identify the Swift executable itself.
-  constexpr static const char * const SWIFT_EXECUTABLE_NAME = "swift";
+  constexpr static const char *const SWIFT_EXECUTABLE_NAME = "swift";
 
   /// Packs together the supplementary information about the job being created.
   class JobContext {
@@ -122,45 +124,36 @@ protected:
 
     InvocationInfo(const char *name, llvm::opt::ArgStringList args = {},
                    decltype(ExtraEnvironment) extraEnv = {})
-      : ExecutableName(name), Arguments(std::move(args)),
-        ExtraEnvironment(std::move(extraEnv)) {}
+        : ExecutableName(name), Arguments(std::move(args)),
+          ExtraEnvironment(std::move(extraEnv)) {}
   };
 
-  virtual InvocationInfo
-  constructInvocation(const CompileJobAction &job,
-                      const JobContext &context) const;
-  virtual InvocationInfo
-  constructInvocation(const InterpretJobAction &job,
-                      const JobContext &context) const;
-  virtual InvocationInfo
-  constructInvocation(const BackendJobAction &job,
-                      const JobContext &context) const;
-  virtual InvocationInfo
-  constructInvocation(const MergeModuleJobAction &job,
-                      const JobContext &context) const;
-  virtual InvocationInfo
-  constructInvocation(const ModuleWrapJobAction &job,
-                      const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const CompileJobAction &job,
+                                             const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const InterpretJobAction &job,
+                                             const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const BackendJobAction &job,
+                                             const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const MergeModuleJobAction &job,
+                                             const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const ModuleWrapJobAction &job,
+                                             const JobContext &context) const;
 
-  virtual InvocationInfo
-  constructInvocation(const REPLJobAction &job,
-                      const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const REPLJobAction &job,
+                                             const JobContext &context) const;
 
-  virtual InvocationInfo
-  constructInvocation(const GenerateDSYMJobAction &job,
-                      const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const GenerateDSYMJobAction &job,
+                                             const JobContext &context) const;
   virtual InvocationInfo
   constructInvocation(const VerifyDebugInfoJobAction &job,
                       const JobContext &context) const;
-  virtual InvocationInfo
-  constructInvocation(const GeneratePCHJobAction &job,
-                      const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const GeneratePCHJobAction &job,
+                                             const JobContext &context) const;
   virtual InvocationInfo
   constructInvocation(const AutolinkExtractJobAction &job,
                       const JobContext &context) const;
-  virtual InvocationInfo
-  constructInvocation(const LinkJobAction &job,
-                      const JobContext &context) const;
+  virtual InvocationInfo constructInvocation(const LinkJobAction &job,
+                                             const JobContext &context) const;
 
   /// Searches for the given executable in appropriate paths relative to the
   /// Swift binary.
@@ -176,6 +169,34 @@ protected:
   /// This method is invoked by findProgramRelativeToSwift().
   virtual std::string findProgramRelativeToSwiftImpl(StringRef name) const;
 
+  void addInputsOfType(llvm::opt::ArgStringList &Arguments,
+                       ArrayRef<const Action *> Inputs,
+                       file_types::ID InputType,
+                       const char *PrefixArgument = nullptr) const;
+
+  void addInputsOfType(llvm::opt::ArgStringList &Arguments,
+                       ArrayRef<const Job *> Jobs,
+                       const llvm::opt::ArgList &Args, file_types::ID InputType,
+                       const char *PrefixArgument = nullptr) const;
+
+  void addPrimaryInputsOfType(llvm::opt::ArgStringList &Arguments,
+                              ArrayRef<const Job *> Jobs,
+                              const llvm::opt::ArgList &Args,
+                              file_types::ID InputType,
+                              const char *PrefixArgument = nullptr) const;
+
+  /// Get the runtime library link path, which is platform-specific and found
+  /// relative to the compiler.
+  void getRuntimeLibraryPath(SmallVectorImpl<char> &runtimeLibPath,
+                             const llvm::opt::ArgList &args, bool shared) const;
+
+  void addPathEnvironmentVariableIfNeeded(Job::EnvironmentVector &env,
+                                          const char *name,
+                                          const char *separator,
+                                          options::ID optionID,
+                                          const llvm::opt::ArgList &args,
+                                          StringRef extraEntry = "") const;
+
 public:
   virtual ~ToolChain() = default;
 
@@ -187,8 +208,7 @@ public:
   ///
   /// This method dispatches to the various \c constructInvocation methods,
   /// which may be overridden by platform-specific subclasses.
-  std::unique_ptr<Job> constructJob(const JobAction &JA,
-                                    Compilation &C,
+  std::unique_ptr<Job> constructJob(const JobAction &JA, Compilation &C,
                                     SmallVectorImpl<const Job *> &&inputs,
                                     ArrayRef<const Action *> inputActions,
                                     std::unique_ptr<CommandOutput> output,
@@ -221,16 +241,35 @@ public:
   /// Return the default language type to use for the given extension.
   /// If the extension is empty or is otherwise not recognized, return
   /// the invalid type \c TY_INVALID.
-  virtual file_types::ID lookupTypeForExtension(StringRef Ext) const;
+  file_types::ID lookupTypeForExtension(StringRef Ext) const;
 
-  /// Check whether a clang library with a given name exists.
+  /// Copies the path for the directory clang libraries would be stored in on
+  /// the current toolchain.
+  void getClangLibraryPath(const llvm::opt::ArgList &Args,
+                           SmallString<128> &LibPath) const;
+
+  /// Returns the name the clang library for a given sanitizer would have on
+  /// the current toolchain.
   ///
-  /// \param args Invocation arguments.
+  /// \param Sanitizer Sanitizer name.
+  /// \param shared Whether the library is shared
+  virtual std::string sanitizerRuntimeLibName(StringRef Sanitizer,
+                                              bool shared = true) const = 0;
+
+  /// Returns whether a given sanitizer exists for the current toolchain.
+  ///
   /// \param sanitizer Sanitizer name.
   /// \param shared Whether the library is shared
-  virtual bool sanitizerRuntimeLibExists(const llvm::opt::ArgList &args,
-                                         StringRef sanitizer,
-                                         bool shared=true) const;
+  bool sanitizerRuntimeLibExists(const llvm::opt::ArgList &args,
+                                 StringRef sanitizer, bool shared = true) const;
+
+  /// Adds a runtime library to the arguments list for linking.
+  ///
+  /// \param LibName The library name
+  /// \param Arguments The arguments list to append to
+  void addLinkRuntimeLib(const llvm::opt::ArgList &Args,
+                         llvm::opt::ArgStringList &Arguments,
+                         StringRef LibName) const;
 };
 } // end namespace driver
 } // end namespace swift
