+++ lib/Driver/ToolChain.cpp
@@ -50,8 +50,7 @@ const char *
 ToolChain::JobContext::getTemporaryFilePath(const llvm::Twine &name,
                                             StringRef suffix) const {
   SmallString<128> buffer;
-  std::error_code EC =
-      llvm::sys::fs::createTemporaryFile(name, suffix, buffer);
+  std::error_code EC = llvm::sys::fs::createTemporaryFile(name, suffix, buffer);
   if (EC) {
     // FIXME: This should not take down the entire process.
     llvm::report_fatal_error("unable to create temporary file for filelist");
@@ -63,13 +62,10 @@ ToolChain::JobContext::getTemporaryFilePath(const llvm::Twine &name,
   return C.getArgs().MakeArgString(buffer.str());
 }
 
-std::unique_ptr<Job>
-ToolChain::constructJob(const JobAction &JA,
-                        Compilation &C,
-                        SmallVectorImpl<const Job *> &&inputs,
-                        ArrayRef<const Action *> inputActions,
-                        std::unique_ptr<CommandOutput> output,
-                        const OutputInfo &OI) const {
+std::unique_ptr<Job> ToolChain::constructJob(
+    const JobAction &JA, Compilation &C, SmallVectorImpl<const Job *> &&inputs,
+    ArrayRef<const Action *> inputActions,
+    std::unique_ptr<CommandOutput> output, const OutputInfo &OI) const {
   JobContext context{C, inputs, inputActions, *output, OI};
 
   auto invocationInfo = [&]() -> InvocationInfo {
@@ -77,17 +73,17 @@ ToolChain::constructJob(const JobAction &JA,
 #define CASE(K)                                                                \
   case Action::Kind::K:                                                        \
     return constructInvocation(cast<K##Action>(JA), context);
-    CASE(CompileJob)
-    CASE(InterpretJob)
-    CASE(BackendJob)
-    CASE(MergeModuleJob)
-    CASE(ModuleWrapJob)
-    CASE(LinkJob)
-    CASE(GenerateDSYMJob)
-    CASE(VerifyDebugInfoJob)
-    CASE(GeneratePCHJob)
-    CASE(AutolinkExtractJob)
-    CASE(REPLJob)
+      CASE(CompileJob)
+      CASE(InterpretJob)
+      CASE(BackendJob)
+      CASE(MergeModuleJob)
+      CASE(ModuleWrapJob)
+      CASE(LinkJob)
+      CASE(GenerateDSYMJob)
+      CASE(VerifyDebugInfoJob)
+      CASE(GeneratePCHJob)
+      CASE(AutolinkExtractJob)
+      CASE(REPLJob)
 #undef CASE
     case Action::Kind::Input:
       llvm_unreachable("not a JobAction");
@@ -153,8 +149,7 @@ file_types::ID ToolChain::lookupTypeForExtension(StringRef Ext) const {
 
 /// Return a _single_ TY_Swift InputAction, if one exists;
 /// if 0 or >1 such inputs exist, return nullptr.
-static const InputAction*
-findSingleSwiftInput(const CompileJobAction *CJA) {
+static const InputAction *findSingleSwiftInput(const CompileJobAction *CJA) {
   auto Inputs = CJA->getInputs();
   const InputAction *IA = nullptr;
   for (auto const *I : Inputs) {
@@ -172,8 +167,7 @@ findSingleSwiftInput(const CompileJobAction *CJA) {
   return IA;
 }
 
-static bool
-jobsHaveSameExecutableNames(const Job *A, const Job *B) {
+static bool jobsHaveSameExecutableNames(const Job *A, const Job *B) {
   // Jobs that get here (that are derived from CompileJobActions) should always
   // have the same executable name -- it should always be SWIFT_EXECUTABLE_NAME
   // -- but we check here just to be sure / fail gracefully in non-assert
@@ -185,16 +179,14 @@ jobsHaveSameExecutableNames(const Job *A, const Job *B) {
   return true;
 }
 
-static bool
-jobsHaveSameOutputTypes(const Job *A, const Job *B) {
+static bool jobsHaveSameOutputTypes(const Job *A, const Job *B) {
   if (A->getOutput().getPrimaryOutputType() !=
       B->getOutput().getPrimaryOutputType())
     return false;
   return A->getOutput().hasSameAdditionalOutputTypes(B->getOutput());
 }
 
-static bool
-jobsHaveSameEnvironment(const Job *A, const Job *B) {
+static bool jobsHaveSameEnvironment(const Job *A, const Job *B) {
   auto AEnv = A->getExtraEnvironment();
   auto BEnv = B->getExtraEnvironment();
   if (AEnv.size() != BEnv.size())
@@ -208,8 +200,7 @@ jobsHaveSameEnvironment(const Job *A, const Job *B) {
   return true;
 }
 
-bool
-ToolChain::jobIsBatchable(const Compilation &C, const Job *A) const {
+bool ToolChain::jobIsBatchable(const Compilation &C, const Job *A) const {
   // FIXME: There might be a tighter criterion to use here?
   if (C.getOutputInfo().CompilerMode != OutputInfo::Mode::StandardCompile)
     return false;
@@ -219,13 +210,11 @@ ToolChain::jobIsBatchable(const Compilation &C, const Job *A) const {
   return findSingleSwiftInput(CJActA) != nullptr;
 }
 
-bool
-ToolChain::jobsAreBatchCombinable(const Compilation &C,
-                                  const Job *A, const Job *B) const {
+bool ToolChain::jobsAreBatchCombinable(const Compilation &C, const Job *A,
+                                       const Job *B) const {
   assert(jobIsBatchable(C, A));
   assert(jobIsBatchable(C, B));
-  return (jobsHaveSameExecutableNames(A, B) &&
-          jobsHaveSameOutputTypes(A, B) &&
+  return (jobsHaveSameExecutableNames(A, B) && jobsHaveSameOutputTypes(A, B) &&
           jobsHaveSameEnvironment(A, B));
 }
 
@@ -286,15 +275,16 @@ mergeBatchInputs(ArrayRef<const Job *> jobs,
 /// discovered yet). So long as this is true, we need to make sure any batch job
 /// we build names its inputs in an order that's a subsequence of the sequence
 /// of inputs the driver was initially invoked with.
-static void sortJobsToMatchCompilationInputs(ArrayRef<const Job *> unsortedJobs,
-                                             SmallVectorImpl<const Job *> &sortedJobs,
-                                             Compilation &C) {
+static void
+sortJobsToMatchCompilationInputs(ArrayRef<const Job *> unsortedJobs,
+                                 SmallVectorImpl<const Job *> &sortedJobs,
+                                 Compilation &C) {
   llvm::StringMap<const Job *> jobsByInput;
   for (const Job *J : unsortedJobs) {
     const CompileJobAction *CJA = cast<CompileJobAction>(&J->getSource());
-    const InputAction* IA = findSingleSwiftInput(CJA);
-    auto R = jobsByInput.insert(std::make_pair(IA->getInputArg().getValue(),
-                                               J));
+    const InputAction *IA = findSingleSwiftInput(CJA);
+    auto R =
+        jobsByInput.insert(std::make_pair(IA->getInputArg().getValue(), J));
     assert(R.second);
   }
   for (const InputPair &P : C.getInputFiles()) {
@@ -310,8 +300,7 @@ static void sortJobsToMatchCompilationInputs(ArrayRef<const Job *> unsortedJobs,
 /// on \p BatchJob, to build the \c InvocationInfo.
 std::unique_ptr<Job>
 ToolChain::constructBatchJob(ArrayRef<const Job *> unsortedJobs,
-                             Compilation &C) const
-{
+                             Compilation &C) const {
   if (unsortedJobs.empty())
     return nullptr;
 
@@ -336,21 +325,9 @@ ToolChain::constructBatchJob(ArrayRef<const Job *> unsortedJobs,
   JobContext context{C, inputJobs.getArrayRef(), inputActions.getArrayRef(),
                      *output, OI};
   auto invocationInfo = constructInvocation(*batchCJA, context);
-  return llvm::make_unique<BatchJob>(*batchCJA,
-                                     inputJobs.takeVector(),
-                                     std::move(output),
-                                     executablePath,
-                                     std::move(invocationInfo.Arguments),
-                                     std::move(invocationInfo.ExtraEnvironment),
-                                     std::move(invocationInfo.FilelistInfos),
-                                     sortedJobs);
-}
-
-bool
-ToolChain::sanitizerRuntimeLibExists(const ArgList &args,
-                                     StringRef sanitizerName,
-                                     bool shared) const {
-  // Assume no sanitizers are supported by default.
-  // This method should be overriden by a platform-specific subclass.
-  return false;
+  return llvm::make_unique<BatchJob>(
+      *batchCJA, inputJobs.takeVector(), std::move(output), executablePath,
+      std::move(invocationInfo.Arguments),
+      std::move(invocationInfo.ExtraEnvironment),
+      std::move(invocationInfo.FilelistInfos), sortedJobs);
 }
